requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
geographiclib==2.0
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.3.0
